moved home page to component
