+++ client/src/App.css
@@ -5,12 +5,3 @@
   text-align: center;
 }
 
-.title {
-  background-color: indianred;
-  height: 10%;
-}
-
-.user {
-  height: 30%;
-  margin-top: 5%;
-}
