+++ client/src/App.tsx
@@ -1,78 +1,10 @@
 import { useEffect, useState } from 'react'
 import './App.css'
 import axios from 'axios'
-import { Avatar, Layout, Menu, MenuProps, Typography } from 'antd'
-import { Content } from 'antd/es/layout/layout'
-import Sider from 'antd/es/layout/Sider'
-import { UserOutlined } from '@ant-design/icons';
-import { AppstoreOutlined, MailOutlined, SettingOutlined } from '@ant-design/icons';
+import HomePage from './pages/HomePage/HomePage'
 
 
 
-const contentStyle: React.CSSProperties = {
-  textAlign: 'center',
-  minHeight: 120,
-  lineHeight: '120px',
-  color: '#fff',
-  backgroundColor: '#0958d9',
-};
-
-const siderStyle: React.CSSProperties = {
-  textAlign: 'center',
-  lineHeight: '120px',
-  color: '#fff',
-  backgroundColor: '#1677ff',
-};
-
-
-const layoutStyle = {
-  overflow: 'auto',
-  width: '100%',
-  maxWidth: '100%',
-  height: '100vh',
-};
-
-type MenuItem = Required<MenuProps>['items'][number];
-
-function getItem(
-  label: React.ReactNode,
-  key: React.Key,
-  icon?: React.ReactNode,
-  children?: MenuItem[],
-  type?: 'group',
-): MenuItem {
-  return {
-    key,
-    icon,
-    children,
-    label,
-    type,
-  } as MenuItem;
-}
-
-const items: MenuProps['items'] = [
-  getItem('Navigation One', 'sub1', <MailOutlined />, [
-    getItem('Item 1', 'g1', null, [getItem('Option 1', '1'), getItem('Option 2', '2')], 'group'),
-    getItem('Item 2', 'g2', null, [getItem('Option 3', '3'), getItem('Option 4', '4')], 'group'),
-  ]),
-
-  getItem('Navigation Two', 'sub2', <AppstoreOutlined />, [
-    getItem('Option 5', '5'),
-    getItem('Option 6', '6'),
-    getItem('Submenu', 'sub3', null, [getItem('Option 7', '7'), getItem('Option 8', '8')]),
-  ]),
-
-  { type: 'divider' },
-
-  getItem('Navigation Three', 'sub4', <SettingOutlined />, [
-    getItem('Option 9', '9'),
-    getItem('Option 10', '10'),
-    getItem('Option 11', '11'),
-    getItem('Option 12', '12'),
-  ]),
-
-  getItem('Group', 'grp', null, [getItem('Option 13', '13'), getItem('Option 14', '14')], 'group'),
-];
 
 function App() {
 
@@ -92,31 +24,7 @@ function App() {
   
   return (
     <div>
-      <Layout style={layoutStyle}>
-        <Layout>
-          <Sider width="20%" style={siderStyle}>
-            <div className="title">
-              <Typography.Title level={2} className='text-center align-middle'>
-                Surge
-              </Typography.Title>
-            </div>
-            <div className="user">
-              <Avatar size={128} icon={<UserOutlined />} />
-              <Typography.Title level={4} style={{ color: 'white' }}>
-                Employee Name
-              </Typography.Title>
-            </div>
-            <Menu
-              style={{ width: 256 }}
-              defaultSelectedKeys={['1']}
-              defaultOpenKeys={['sub1']}
-              mode="inline"
-              items={items}
-            />
-          </Sider>
-          <Content style={contentStyle}>Content</Content>
-        </Layout>
-      </Layout>
+      <HomePage />
     </div>
   )
 }
