+++ client/src/pages/HomePage/HomePage.css
@@ -1,0 +1,9 @@
+.title {
+    background-color: indianred;
+    height: 10%;
+  }
+  
+  .user {
+    height: 30%;
+    margin-top: 5%;
+  }
